Transfers code from form-switcher to form file
